# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stock Management",
        "description": "API for managing stock items, categories, and sales",
        "contact": {
            "email": "inolatse@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/sell/article": {
            "post": {
                "tags": [
                    "OutStore"
                ],
                "summary": "Create a single OutStore entry",
                "description": "Enregistre une seule sortie d'article du stock.",
                "operationId": "280e03da220aa42ce8b4a496ecdfa779",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "article_id"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "article_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article removed from stock successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OutStore"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/sell/articles": {
            "post": {
                "tags": [
                    "OutStore"
                ],
                "summary": "Create multiple OutStore entries",
                "description": "Enregistre plusieurs sorties d'articles du stock en une seule requête.",
                "operationId": "c8fa4574f6b328efd3264f56088aaec2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "quantity",
                                                "article_id"
                                            ],
                                            "properties": {
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "article_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Articles removed from stock successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorties de stock créées avec succès."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OutStore"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/v1/sell/articles/statistics/period": {
            "get": {
                "tags": [
                    "OutStore"
                ],
                "summary": "Retrieve statistics on article sales",
                "description": "Obtient des statistiques sur les sorties d'articles, comme la quantité totale, les articles les plus sortis et les sorties dans une plage de dates.",
                "operationId": "af6c8da6b065b46eda1f5085650554a8",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Date de début pour filtrer les statistiques (format: YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Date de fin pour filtrer les statistiques (format: YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_quantity": {
                                            "description": "Quantité totale des articles sortis",
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "articles": {
                                            "description": "Liste des articles avec leurs statistiques",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "article_id": {
                                                        "description": "ID de l'article",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "article_name": {
                                                        "description": "Nom de l'article",
                                                        "type": "string",
                                                        "example": "Article A"
                                                    },
                                                    "quantity": {
                                                        "description": "Quantité totale sortie pour cet article",
                                                        "type": "integer",
                                                        "example": 50
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "most_sold": {
                                            "description": "Article le plus vendu dans la période sélectionnée",
                                            "properties": {
                                                "article_id": {
                                                    "description": "ID de l'article",
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "article_name": {
                                                    "description": "Nom de l'article",
                                                    "type": "string",
                                                    "example": "Article C"
                                                },
                                                "quantity": {
                                                    "description": "Quantité totale sortie pour cet article",
                                                    "type": "integer",
                                                    "example": 70
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide (exemple : format de date incorrect)"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "OutStore": {
                "required": [
                    "quantity",
                    "solde",
                    "soldeTotal",
                    "article_id"
                ],
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float"
                    },
                    "soldeTotal": {
                        "type": "number",
                        "format": "float"
                    },
                    "article_id": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            }
        }
    },
...
}